Removed divider row component
