+++ apps/scoutgame/components/claim/components/PointsTable/PointsReceiptRewardRow.tsx
@@ -8,7 +8,6 @@ import type {
   SoldNftsPointsReceiptReward
 } from 'lib/points/getPointsReceiptsRewards';
 
-import { DividerRow } from '../common/DividerRow';
 import { PointsCell } from '../common/PointsCell';
 
 function getOrdinal(n: number): string {
@@ -19,20 +18,17 @@ function getOrdinal(n: number): string {
 
 function BuilderRewardRow({ builderReward }: { builderReward: BuilderPointsReceiptReward }) {
   return (
-    <>
-      <DividerRow />
-      <TableRow key={`${builderReward.period}builder-rewards`}>
-        <TableCell align='left'>
-          <Typography>Builder rewards</Typography>
-        </TableCell>
-        <TableCell align='center'>
-          <Typography>{builderReward.period}</Typography>
-        </TableCell>
-        <TableCell align='right'>
-          <PointsCell points={builderReward.points} />
-        </TableCell>
-      </TableRow>
-    </>
+    <TableRow>
+      <TableCell align='left'>
+        <Typography>Builder rewards</Typography>
+      </TableCell>
+      <TableCell align='center'>
+        <Typography>{builderReward.period}</Typography>
+      </TableCell>
+      <TableCell align='right'>
+        <PointsCell points={builderReward.points} />
+      </TableCell>
+    </TableRow>
   );
 }
 
@@ -42,42 +38,36 @@ function LeaderboardRankRewardRow({
   leaderboardRankReward: LeaderboardRankPointsReceiptReward;
 }) {
   return (
-    <>
-      <DividerRow />
-      <TableRow key={`${leaderboardRankReward.period}rank`}>
-        <TableCell align='left'>
-          <Typography>Finished {getOrdinal(leaderboardRankReward.rank)}</Typography>
-        </TableCell>
-        <TableCell align='center'>
-          <Typography>{leaderboardRankReward.period}</Typography>
-        </TableCell>
-        <TableCell align='right'>
-          <PointsCell points={leaderboardRankReward.points} />
-        </TableCell>
-      </TableRow>
-    </>
+    <TableRow>
+      <TableCell align='left'>
+        <Typography>Finished {getOrdinal(leaderboardRankReward.rank)}</Typography>
+      </TableCell>
+      <TableCell align='center'>
+        <Typography>{leaderboardRankReward.period}</Typography>
+      </TableCell>
+      <TableCell align='right'>
+        <PointsCell points={leaderboardRankReward.points} />
+      </TableCell>
+    </TableRow>
   );
 }
 
 function SoldNftsRewardRow({ soldNftsReward }: { soldNftsReward: SoldNftsPointsReceiptReward }) {
   return (
-    <>
-      <DividerRow />
-      <TableRow key={`${soldNftsReward.period}sold-nft`}>
-        <TableCell align='left'>
-          <Stack direction='row' alignItems='center' justifyContent='flex-start' gap={0.5}>
-            <Typography>Sold {soldNftsReward.quantity}</Typography>
-            <Image alt='card' src='/images/profile/icons/card.svg' width={18} height={18} />
-          </Stack>
-        </TableCell>
-        <TableCell align='center'>
-          <Typography>{soldNftsReward.period}</Typography>
-        </TableCell>
-        <TableCell align='right'>
-          <PointsCell points={soldNftsReward.points} />
-        </TableCell>
-      </TableRow>
-    </>
+    <TableRow>
+      <TableCell align='left'>
+        <Stack direction='row' alignItems='center' justifyContent='flex-start' gap={0.5}>
+          <Typography>Sold {soldNftsReward.quantity}</Typography>
+          <Image alt='card' src='/images/profile/icons/card.svg' width={18} height={18} />
+        </Stack>
+      </TableCell>
+      <TableCell align='center'>
+        <Typography>{soldNftsReward.period}</Typography>
+      </TableCell>
+      <TableCell align='right'>
+        <PointsCell points={soldNftsReward.points} />
+      </TableCell>
+    </TableRow>
   );
 }
